Added authorization with Gate for edit/delete question

+++ app/Http/Controllers/QuestionsController.php
@@ -75,6 +75,19 @@ class QuestionsController extends Controller
     public function edit(Question $question)
     {
         //$question = Question::findOrFail($id);
+        
+        /*
+        if(\Gate::allows('update-question', $question))
+        {        
+            return view("questions.edit", compact('question'));
+        }
+        abort(403, "Access denied");
+        */
+        
+        if(\Gate::denies('update-question', $question))
+        {
+            abort(403, "Access denied");
+        }
         return view("questions.edit", compact('question'));
     }
 
@@ -87,6 +100,11 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if(\Gate::denies('update-question', $question))
+        {
+            abort(403, "Access denied");
+        }
+        
         $question->update($request->only('title', 'body'));
         
         return redirect()->route('questions.index')->with('success', 'Your Question has been updated');
@@ -100,6 +118,12 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        
+        if(\Gate::denies('delete-question', $question))
+        {
+            abort(403, "Access denied");
+        }
+        
         $question->delete();
         
         return redirect()->route('questions.index')->with('success', 'Your Question has been deleted');
